requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.20.0
kaleido>=0.2.1
openpyxl>=3.1.2
gspread>=6.0.0
oauth2client>=4.1.3
jinja2>=3.1.3
weasyprint>=61.0
python-dotenv>=1.0.1
pytest>=8.0.0
chardet>=5.2.0
scipy>=1.12.0
